refactor(ChallengeList): challenge card
